refactor(fxspot): polish slice contracts and response naming
