+++ backend/src/main/java/com/alligator/market/backend/instrument/asset/forex/fxspot/api/command/update/mapper/UpdateFxSpotRequestMapper.java
@@ -1,7 +1,7 @@
 package com.alligator.market.backend.instrument.asset.forex.fxspot.api.command.update.mapper;
 
 import com.alligator.market.backend.instrument.asset.forex.fxspot.api.command.update.dto.UpdateFxSpotRequest;
-import com.alligator.market.backend.instrument.asset.forex.fxspot.application.command.update.dto.UpdateFxSpotCommand;
+import com.alligator.market.backend.instrument.asset.forex.fxspot.application.command.update.UpdateFxSpotCommand;
 import com.alligator.market.domain.instrument.vo.InstrumentCode;
 
 import java.util.Objects;
