+++ backend/src/main/java/com/alligator/market/backend/instrument/asset/forex/fxspot/application/command/create/CreateFxSpotCommand.java
@@ -1,4 +1,4 @@
-package com.alligator.market.backend.instrument.asset.forex.fxspot.application.command.create.dto;
+package com.alligator.market.backend.instrument.asset.forex.fxspot.application.command.create;
 
 import com.alligator.market.domain.instrument.asset.forex.fxspot.classification.FxSpotTenor;
